+++ db-metadata-web/package/utils/dialog.js
@@ -82,7 +82,7 @@ export function createDialog(Vue) {
                             return this.meta.conf
                         },
                         width() {
-                            const {meta: {style: {width: widthInMeta = '50%'} = {}} = {}, conf: {width: widthInAttr}} = this
+                            const {meta: {style: {width: widthInMeta} = {}} = {}, conf: {width: widthInAttr = '50%'}} = this
                             return utils.assertEmpty(widthInMeta, widthInAttr)
                         }
                     }